초 (seconds) 등 좌표는 float 로 통일.

+++ Rhythm/Rhythm/Note.h
@@ -7,7 +7,7 @@ public:
 	};
 	int noteType;
 	int position; // 1=s, 2=d, 3=f, 4=j, 5=k, 6=l
-	double seconds;
+	float seconds;
 
 	Note();
 	virtual ~Note();
